created first version with hash

+++ src/globals.h
@@ -34,7 +34,7 @@ extern uint64_t uptimecounter;
 extern osjob_t sendjob;
 extern int macnum, blenum, countermode, screensaver, adrmode, lorasf, txpower, rlim;
 extern bool joinstate;
-extern std::set<uint64_t> macs;
+extern std::set<uint32_t> macs; 
 
 #ifdef HAS_DISPLAY
     extern HAS_DISPLAY u8x8;
